moves farms to cluster
