+++ resources/views/filament/app/resources/farm-resource/widgets/farm-list-header-widget.blade.php
@@ -2,7 +2,7 @@
     @if(!$this->hasLocations())
         <x-filament::section title="Locations" icon="heroicon-o-exclamation-circle" icon-color="danger" heading="No Viable Locations Found">
             <p>
-                There are no locations that can have farms assigned. Please check the <a class="underline text-blue-600" href="{{\App\Filament\App\Resources\LocationLevelResource::getUrl('index')}}">sample frame page</a> to add administrative levels and locations before you import your list of farms.
+                There are no locations that can have farms assigned. Please check the <a class="underline text-blue-600" href="{{\App\Filament\App\Clusters\LocationLevels\Resources\LocationLevelResource::getUrl('index')}}">sample frame page</a> to add administrative levels and locations before you import your list of farms.
 
             </p>
         </x-filament::section>
